% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_csv}
\alias{save_csv}
\title{Write (new) csv and append if it already exists}
\usage{
save_csv(d, path, create_subfolders = T)
}
\arguments{
\item{d}{The dataset you want to write}

\item{path}{The path you want to save the csv to}

\item{create_subfolders}{defaults to \code{TRUE}. Will create any subfolders you specify in \code{path}.}
}
\description{
This function writes a csv, creates the necessary (sub-)folders and appends the .csv data if it already is present.
\cr \cr Be aware this WILL create new subfolders by default if you specify folders that don't exist. If you don't want this behavior turn it off with \code{create_subfolders = FALSE}.
}
\examples{
\dontrun{
if (interactive()) {
  ## write the csv (or append if it already exists)
  save_csv(cars, "cars.csv")

  ## create a folder called "data" (if it doesn't exist yet) before it saves the csv
  save_csv(cars, "data/cars.csv")
}
}
}
