% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save_lines}
\alias{save_lines}
\title{Write (new) lines and append if it already exists}
\usage{
save_lines(l, path, create_subfolders)
}
\arguments{
\item{l}{The lines you want to write}

\item{path}{The path you want to save the csv to}

\item{create_subfolders}{defaults to \code{TRUE}. Will create any subfolders you specify in \code{path}.}
}
\description{
This function writes a text file, creates the necessary (sub-)folders and appends the .csv data if it already is present.
\cr \cr Be aware this WILL create new subfolders by default if you specify folders that don't exist. If you don't want this behavior turn it off with \code{create_subfolders = FALSE}.
}
\examples{
\dontrun{
if (interactive()) {
  ## write the lines (or append if it already exists)
  save_lines(names(mtcars), "carnames.txt")

  ## create a folder called "txt" (if it doesn't exist yet) before it saves the lines
  save_lines(names(mtcars), "txt/carnames.txt")
}
}
}
